Initial commit
This repository is a sort of sketch pad for exploring various approaches
to deploying kubernetes. At least for the time being, we're targeting a
homelab setup with fairly minimal hardware (a lenovo m920q mini pc with
32Gb of ram). The hardware runs proxmox, so all the scripts aim to adapt
themselves to that cloud provider.

This commit includes an initial vanilla setup from chapter 06 of
the book of kubernetes. The ansible scripts from the book have been
modified slightly to run on the homelab proxmox setup. The *.tf
files included in the infra/ directory enable the deployment of the
virtual machines required to run the cluster. The biggest stumbling
block during the translation from aws to proxmox was failure to
recognize the need to define and attach a second network interface to
all the VMs. The IP for this interface is what is used for internal
communication between the VMs, and it is defined as the ansible variable
"k8s_control_plane_endpoint".